+++ test/e2e/common/downwardapi_volume.go
@@ -21,7 +21,6 @@ import (
 	"time"
 
 	"k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/uuid"
 	"k8s.io/kubernetes/test/e2e/framework"
@@ -30,7 +29,7 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-var _ = framework.KubeDescribe("Downward API volume", func() {
+var _ = Describe("[sig-storage] Downward API volume", func() {
 	// How long to wait for a log pod to be displayed
 	const podLogTimeout = 2 * time.Minute
 	f := framework.NewDefaultFramework("downward-api")
@@ -39,7 +38,7 @@ var _ = framework.KubeDescribe("Downward API volume", func() {
 		podClient = f.PodClient()
 	})
 
-	It("should provide podname only [Conformance] [sig-storage]", func() {
+	It("should provide podname only [Conformance]", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumePodForSimpleTest(podName, "/etc/podname")
 
@@ -48,7 +47,7 @@ var _ = framework.KubeDescribe("Downward API volume", func() {
 		})
 	})
 
-	It("should set DefaultMode on files [Conformance] [sig-storage]", func() {
+	It("should set DefaultMode on files [Conformance]", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		defaultMode := int32(0400)
 		pod := downwardAPIVolumePodForModeTest(podName, "/etc/podname", nil, &defaultMode)
@@ -58,7 +57,7 @@ var _ = framework.KubeDescribe("Downward API volume", func() {
 		})
 	})
 
-	It("should set mode on item file [Conformance] [sig-storage]", func() {
+	It("should set mode on item file [Conformance]", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		mode := int32(0400)
 		pod := downwardAPIVolumePodForModeTest(podName, "/etc/podname", &mode, nil)
@@ -68,7 +67,7 @@ var _ = framework.KubeDescribe("Downward API volume", func() {
 		})
 	})
 
-	It("should provide podname as non-root with fsgroup [Feature:FSGroup] [sig-storage]", func() {
+	It("should provide podname as non-root with fsgroup [Feature:FSGroup]", func() {
 		podName := "metadata-volume-" + string(uuid.NewUUID())
 		uid := int64(1001)
 		gid := int64(1234)
@@ -82,7 +81,7 @@ var _ = framework.KubeDescribe("Downward API volume", func() {
 		})
 	})
 
-	It("should provide podname as non-root with fsgroup and defaultMode [Feature:FSGroup] [sig-storage]", func() {
+	It("should provide podname as non-root with fsgroup and defaultMode [Feature:FSGroup]", func() {
 		podName := "metadata-volume-" + string(uuid.NewUUID())
 		uid := int64(1001)
 		gid := int64(1234)
@@ -97,7 +96,7 @@ var _ = framework.KubeDescribe("Downward API volume", func() {
 		})
 	})
 
-	It("should update labels on modification [Conformance] [sig-storage]", func() {
+	It("should update labels on modification [Conformance]", func() {
 		labels := map[string]string{}
 		labels["key1"] = "value1"
 		labels["key2"] = "value2"
@@ -124,7 +123,7 @@ var _ = framework.KubeDescribe("Downward API volume", func() {
 			podLogTimeout, framework.Poll).Should(ContainSubstring("key3=\"value3\"\n"))
 	})
 
-	It("should update annotations on modification [Conformance] [sig-storage]", func() {
+	It("should update annotations on modification [Conformance]", func() {
 		annotations := map[string]string{}
 		annotations["builder"] = "bar"
 		podName := "annotationupdate" + string(uuid.NewUUID())
@@ -153,7 +152,7 @@ var _ = framework.KubeDescribe("Downward API volume", func() {
 			podLogTimeout, framework.Poll).Should(ContainSubstring("builder=\"foo\"\n"))
 	})
 
-	It("should provide container's cpu limit [Conformance] [sig-storage]", func() {
+	It("should provide container's cpu limit [Conformance]", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumeForContainerResources(podName, "/etc/cpu_limit")
 
@@ -162,7 +161,7 @@ var _ = framework.KubeDescribe("Downward API volume", func() {
 		})
 	})
 
-	It("should provide container's memory limit [Conformance] [sig-storage]", func() {
+	It("should provide container's memory limit [Conformance]", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumeForContainerResources(podName, "/etc/memory_limit")
 
@@ -171,7 +170,7 @@ var _ = framework.KubeDescribe("Downward API volume", func() {
 		})
 	})
 
-	It("should provide container's cpu request [Conformance] [sig-storage]", func() {
+	It("should provide container's cpu request [Conformance]", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumeForContainerResources(podName, "/etc/cpu_request")
 
@@ -180,7 +179,7 @@ var _ = framework.KubeDescribe("Downward API volume", func() {
 		})
 	})
 
-	It("should provide container's memory request [Conformance] [sig-storage]", func() {
+	It("should provide container's memory request [Conformance]", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumeForContainerResources(podName, "/etc/memory_request")
 
@@ -189,14 +188,14 @@ var _ = framework.KubeDescribe("Downward API volume", func() {
 		})
 	})
 
-	It("should provide node allocatable (cpu) as default cpu limit if the limit is not set [Conformance] [sig-storage]", func() {
+	It("should provide node allocatable (cpu) as default cpu limit if the limit is not set [Conformance]", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumeForDefaultContainerResources(podName, "/etc/cpu_limit")
 
 		f.TestContainerOutputRegexp("downward API volume plugin", pod, 0, []string{"[1-9]"})
 	})
 
-	It("should provide node allocatable (memory) as default memory limit if the limit is not set [Conformance] [sig-storage]", func() {
+	It("should provide node allocatable (memory) as default memory limit if the limit is not set [Conformance]", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumeForDefaultContainerResources(podName, "/etc/memory_limit")
 
@@ -204,214 +203,3 @@ var _ = framework.KubeDescribe("Downward API volume", func() {
 	})
 
 })
-
-func downwardAPIVolumePodForModeTest(name, filePath string, itemMode, defaultMode *int32) *v1.Pod {
-	pod := downwardAPIVolumeBasePod(name, nil, nil)
-
-	pod.Spec.Containers = []v1.Container{
-		{
-			Name:    "client-container",
-			Image:   "gcr.io/google_containers/mounttest:0.8",
-			Command: []string{"/mt", "--file_mode=" + filePath},
-			VolumeMounts: []v1.VolumeMount{
-				{
-					Name:      "podinfo",
-					MountPath: "/etc",
-				},
-			},
-		},
-	}
-	if itemMode != nil {
-		pod.Spec.Volumes[0].VolumeSource.DownwardAPI.Items[0].Mode = itemMode
-	}
-	if defaultMode != nil {
-		pod.Spec.Volumes[0].VolumeSource.DownwardAPI.DefaultMode = defaultMode
-	}
-
-	return pod
-}
-
-func downwardAPIVolumePodForSimpleTest(name string, filePath string) *v1.Pod {
-	pod := downwardAPIVolumeBasePod(name, nil, nil)
-
-	pod.Spec.Containers = []v1.Container{
-		{
-			Name:    "client-container",
-			Image:   "gcr.io/google_containers/mounttest:0.8",
-			Command: []string{"/mt", "--file_content=" + filePath},
-			VolumeMounts: []v1.VolumeMount{
-				{
-					Name:      "podinfo",
-					MountPath: "/etc",
-					ReadOnly:  false,
-				},
-			},
-		},
-	}
-
-	return pod
-}
-
-func downwardAPIVolumeForContainerResources(name string, filePath string) *v1.Pod {
-	pod := downwardAPIVolumeBasePod(name, nil, nil)
-	pod.Spec.Containers = downwardAPIVolumeBaseContainers("client-container", filePath)
-	return pod
-}
-
-func downwardAPIVolumeForDefaultContainerResources(name string, filePath string) *v1.Pod {
-	pod := downwardAPIVolumeBasePod(name, nil, nil)
-	pod.Spec.Containers = downwardAPIVolumeDefaultBaseContainer("client-container", filePath)
-	return pod
-}
-
-func downwardAPIVolumeBaseContainers(name, filePath string) []v1.Container {
-	return []v1.Container{
-		{
-			Name:    name,
-			Image:   "gcr.io/google_containers/mounttest:0.8",
-			Command: []string{"/mt", "--file_content=" + filePath},
-			Resources: v1.ResourceRequirements{
-				Requests: v1.ResourceList{
-					v1.ResourceCPU:    resource.MustParse("250m"),
-					v1.ResourceMemory: resource.MustParse("32Mi"),
-				},
-				Limits: v1.ResourceList{
-					v1.ResourceCPU:    resource.MustParse("1250m"),
-					v1.ResourceMemory: resource.MustParse("64Mi"),
-				},
-			},
-			VolumeMounts: []v1.VolumeMount{
-				{
-					Name:      "podinfo",
-					MountPath: "/etc",
-					ReadOnly:  false,
-				},
-			},
-		},
-	}
-
-}
-
-func downwardAPIVolumeDefaultBaseContainer(name, filePath string) []v1.Container {
-	return []v1.Container{
-		{
-			Name:    name,
-			Image:   "gcr.io/google_containers/mounttest:0.8",
-			Command: []string{"/mt", "--file_content=" + filePath},
-			VolumeMounts: []v1.VolumeMount{
-				{
-					Name:      "podinfo",
-					MountPath: "/etc",
-				},
-			},
-		},
-	}
-
-}
-
-func downwardAPIVolumePodForUpdateTest(name string, labels, annotations map[string]string, filePath string) *v1.Pod {
-	pod := downwardAPIVolumeBasePod(name, labels, annotations)
-
-	pod.Spec.Containers = []v1.Container{
-		{
-			Name:    "client-container",
-			Image:   "gcr.io/google_containers/mounttest:0.8",
-			Command: []string{"/mt", "--break_on_expected_content=false", "--retry_time=120", "--file_content_in_loop=" + filePath},
-			VolumeMounts: []v1.VolumeMount{
-				{
-					Name:      "podinfo",
-					MountPath: "/etc",
-					ReadOnly:  false,
-				},
-			},
-		},
-	}
-
-	applyLabelsAndAnnotationsToDownwardAPIPod(labels, annotations, pod)
-	return pod
-}
-
-func downwardAPIVolumeBasePod(name string, labels, annotations map[string]string) *v1.Pod {
-	pod := &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        name,
-			Labels:      labels,
-			Annotations: annotations,
-		},
-		Spec: v1.PodSpec{
-			Volumes: []v1.Volume{
-				{
-					Name: "podinfo",
-					VolumeSource: v1.VolumeSource{
-						DownwardAPI: &v1.DownwardAPIVolumeSource{
-							Items: []v1.DownwardAPIVolumeFile{
-								{
-									Path: "podname",
-									FieldRef: &v1.ObjectFieldSelector{
-										APIVersion: "v1",
-										FieldPath:  "metadata.name",
-									},
-								},
-								{
-									Path: "cpu_limit",
-									ResourceFieldRef: &v1.ResourceFieldSelector{
-										ContainerName: "client-container",
-										Resource:      "limits.cpu",
-									},
-								},
-								{
-									Path: "cpu_request",
-									ResourceFieldRef: &v1.ResourceFieldSelector{
-										ContainerName: "client-container",
-										Resource:      "requests.cpu",
-									},
-								},
-								{
-									Path: "memory_limit",
-									ResourceFieldRef: &v1.ResourceFieldSelector{
-										ContainerName: "client-container",
-										Resource:      "limits.memory",
-									},
-								},
-								{
-									Path: "memory_request",
-									ResourceFieldRef: &v1.ResourceFieldSelector{
-										ContainerName: "client-container",
-										Resource:      "requests.memory",
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			RestartPolicy: v1.RestartPolicyNever,
-		},
-	}
-
-	return pod
-}
-
-func applyLabelsAndAnnotationsToDownwardAPIPod(labels, annotations map[string]string, pod *v1.Pod) {
-	if len(labels) > 0 {
-		pod.Spec.Volumes[0].DownwardAPI.Items = append(pod.Spec.Volumes[0].DownwardAPI.Items, v1.DownwardAPIVolumeFile{
-			Path: "labels",
-			FieldRef: &v1.ObjectFieldSelector{
-				APIVersion: "v1",
-				FieldPath:  "metadata.labels",
-			},
-		})
-	}
-
-	if len(annotations) > 0 {
-		pod.Spec.Volumes[0].DownwardAPI.Items = append(pod.Spec.Volumes[0].DownwardAPI.Items, v1.DownwardAPIVolumeFile{
-			Path: "annotations",
-			FieldRef: &v1.ObjectFieldSelector{
-				APIVersion: "v1",
-				FieldPath:  "metadata.annotations",
-			},
-		})
-	}
-}
-
-// TODO: add test-webserver example as pointed out in https://github.com/kubernetes/kubernetes/pull/5093#discussion-diff-37606771
