+++ test/e2e/common/secrets.go
@@ -19,7 +19,6 @@ package common
 import (
 	"fmt"
 	"os"
-	"path"
 
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -27,302 +26,10 @@ import (
 	"k8s.io/kubernetes/test/e2e/framework"
 
 	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
 )
 
-var _ = framework.KubeDescribe("Secrets", func() {
+var _ = Describe("[sig-api-machinery] Secrets", func() {
 	f := framework.NewDefaultFramework("secrets")
-
-	It("should be consumable from pods in volume [Conformance] [sig-storage]", func() {
-		doSecretE2EWithoutMapping(f, nil /* default mode */, "secret-test-"+string(uuid.NewUUID()), nil, nil)
-	})
-
-	It("should be consumable from pods in volume with defaultMode set [Conformance] [sig-storage]", func() {
-		defaultMode := int32(0400)
-		doSecretE2EWithoutMapping(f, &defaultMode, "secret-test-"+string(uuid.NewUUID()), nil, nil)
-	})
-
-	It("should be consumable from pods in volume as non-root with defaultMode and fsGroup set [Conformance] [sig-storage]", func() {
-		defaultMode := int32(0440) /* setting fsGroup sets mode to at least 440 */
-		fsGroup := int64(1001)
-		uid := int64(1000)
-		doSecretE2EWithoutMapping(f, &defaultMode, "secret-test-"+string(uuid.NewUUID()), &fsGroup, &uid)
-	})
-
-	It("should be consumable from pods in volume with mappings [Conformance] [sig-storage]", func() {
-		doSecretE2EWithMapping(f, nil)
-	})
-
-	It("should be consumable from pods in volume with mappings and Item Mode set [Conformance] [sig-storage]", func() {
-		mode := int32(0400)
-		doSecretE2EWithMapping(f, &mode)
-	})
-
-	It("should be able to mount in a volume regardless of a different secret existing with same name in different namespace [sig-storage]", func() {
-		var (
-			namespace2  *v1.Namespace
-			err         error
-			secret2Name = "secret-test-" + string(uuid.NewUUID())
-		)
-
-		if namespace2, err = f.CreateNamespace("secret-namespace", nil); err != nil {
-			framework.Failf("unable to create new namespace %s: %v", namespace2.Name, err)
-		}
-
-		secret2 := secretForTest(namespace2.Name, secret2Name)
-		secret2.Data = map[string][]byte{
-			"this_should_not_match_content_of_other_secret": []byte("similarly_this_should_not_match_content_of_other_secret\n"),
-		}
-		if secret2, err = f.ClientSet.Core().Secrets(namespace2.Name).Create(secret2); err != nil {
-			framework.Failf("unable to create test secret %s: %v", secret2.Name, err)
-		}
-		doSecretE2EWithoutMapping(f, nil /* default mode */, secret2.Name, nil, nil)
-	})
-
-	It("should be consumable in multiple volumes in a pod [Conformance] [sig-storage]", func() {
-		// This test ensures that the same secret can be mounted in multiple
-		// volumes in the same pod.  This test case exists to prevent
-		// regressions that break this use-case.
-		var (
-			name             = "secret-test-" + string(uuid.NewUUID())
-			volumeName       = "secret-volume"
-			volumeMountPath  = "/etc/secret-volume"
-			volumeName2      = "secret-volume-2"
-			volumeMountPath2 = "/etc/secret-volume-2"
-			secret           = secretForTest(f.Namespace.Name, name)
-		)
-
-		By(fmt.Sprintf("Creating secret with name %s", secret.Name))
-		var err error
-		if secret, err = f.ClientSet.Core().Secrets(f.Namespace.Name).Create(secret); err != nil {
-			framework.Failf("unable to create test secret %s: %v", secret.Name, err)
-		}
-
-		pod := &v1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "pod-secrets-" + string(uuid.NewUUID()),
-			},
-			Spec: v1.PodSpec{
-				Volumes: []v1.Volume{
-					{
-						Name: volumeName,
-						VolumeSource: v1.VolumeSource{
-							Secret: &v1.SecretVolumeSource{
-								SecretName: name,
-							},
-						},
-					},
-					{
-						Name: volumeName2,
-						VolumeSource: v1.VolumeSource{
-							Secret: &v1.SecretVolumeSource{
-								SecretName: name,
-							},
-						},
-					},
-				},
-				Containers: []v1.Container{
-					{
-						Name:  "secret-volume-test",
-						Image: "gcr.io/google_containers/mounttest:0.8",
-						Args: []string{
-							"--file_content=/etc/secret-volume/data-1",
-							"--file_mode=/etc/secret-volume/data-1"},
-						VolumeMounts: []v1.VolumeMount{
-							{
-								Name:      volumeName,
-								MountPath: volumeMountPath,
-								ReadOnly:  true,
-							},
-							{
-								Name:      volumeName2,
-								MountPath: volumeMountPath2,
-								ReadOnly:  true,
-							},
-						},
-					},
-				},
-				RestartPolicy: v1.RestartPolicyNever,
-			},
-		}
-
-		f.TestContainerOutput("consume secrets", pod, 0, []string{
-			"content of file \"/etc/secret-volume/data-1\": value-1",
-			"mode of file \"/etc/secret-volume/data-1\": -rw-r--r--",
-		})
-	})
-
-	It("optional updates should be reflected in volume [Conformance] [sig-storage]", func() {
-		podLogTimeout := framework.GetPodSecretUpdateTimeout(f.ClientSet)
-		containerTimeoutArg := fmt.Sprintf("--retry_time=%v", int(podLogTimeout.Seconds()))
-		trueVal := true
-		volumeMountPath := "/etc/secret-volumes"
-
-		deleteName := "s-test-opt-del-" + string(uuid.NewUUID())
-		deleteContainerName := "dels-volume-test"
-		deleteVolumeName := "deletes-volume"
-		deleteSecret := &v1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: f.Namespace.Name,
-				Name:      deleteName,
-			},
-			Data: map[string][]byte{
-				"data-1": []byte("value-1"),
-			},
-		}
-
-		updateName := "s-test-opt-upd-" + string(uuid.NewUUID())
-		updateContainerName := "upds-volume-test"
-		updateVolumeName := "updates-volume"
-		updateSecret := &v1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: f.Namespace.Name,
-				Name:      updateName,
-			},
-			Data: map[string][]byte{
-				"data-1": []byte("value-1"),
-			},
-		}
-
-		createName := "s-test-opt-create-" + string(uuid.NewUUID())
-		createContainerName := "creates-volume-test"
-		createVolumeName := "creates-volume"
-		createSecret := &v1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: f.Namespace.Name,
-				Name:      createName,
-			},
-			Data: map[string][]byte{
-				"data-1": []byte("value-1"),
-			},
-		}
-
-		By(fmt.Sprintf("Creating secret with name %s", deleteSecret.Name))
-		var err error
-		if deleteSecret, err = f.ClientSet.Core().Secrets(f.Namespace.Name).Create(deleteSecret); err != nil {
-			framework.Failf("unable to create test secret %s: %v", deleteSecret.Name, err)
-		}
-
-		By(fmt.Sprintf("Creating secret with name %s", updateSecret.Name))
-		if updateSecret, err = f.ClientSet.Core().Secrets(f.Namespace.Name).Create(updateSecret); err != nil {
-			framework.Failf("unable to create test secret %s: %v", updateSecret.Name, err)
-		}
-
-		pod := &v1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "pod-secrets-" + string(uuid.NewUUID()),
-			},
-			Spec: v1.PodSpec{
-				Volumes: []v1.Volume{
-					{
-						Name: deleteVolumeName,
-						VolumeSource: v1.VolumeSource{
-							Secret: &v1.SecretVolumeSource{
-								SecretName: deleteName,
-								Optional:   &trueVal,
-							},
-						},
-					},
-					{
-						Name: updateVolumeName,
-						VolumeSource: v1.VolumeSource{
-							Secret: &v1.SecretVolumeSource{
-								SecretName: updateName,
-								Optional:   &trueVal,
-							},
-						},
-					},
-					{
-						Name: createVolumeName,
-						VolumeSource: v1.VolumeSource{
-							Secret: &v1.SecretVolumeSource{
-								SecretName: createName,
-								Optional:   &trueVal,
-							},
-						},
-					},
-				},
-				Containers: []v1.Container{
-					{
-						Name:    deleteContainerName,
-						Image:   "gcr.io/google_containers/mounttest:0.8",
-						Command: []string{"/mt", "--break_on_expected_content=false", containerTimeoutArg, "--file_content_in_loop=/etc/secret-volumes/delete/data-1"},
-						VolumeMounts: []v1.VolumeMount{
-							{
-								Name:      deleteVolumeName,
-								MountPath: path.Join(volumeMountPath, "delete"),
-								ReadOnly:  true,
-							},
-						},
-					},
-					{
-						Name:    updateContainerName,
-						Image:   "gcr.io/google_containers/mounttest:0.8",
-						Command: []string{"/mt", "--break_on_expected_content=false", containerTimeoutArg, "--file_content_in_loop=/etc/secret-volumes/update/data-3"},
-						VolumeMounts: []v1.VolumeMount{
-							{
-								Name:      updateVolumeName,
-								MountPath: path.Join(volumeMountPath, "update"),
-								ReadOnly:  true,
-							},
-						},
-					},
-					{
-						Name:    createContainerName,
-						Image:   "gcr.io/google_containers/mounttest:0.8",
-						Command: []string{"/mt", "--break_on_expected_content=false", containerTimeoutArg, "--file_content_in_loop=/etc/secret-volumes/create/data-1"},
-						VolumeMounts: []v1.VolumeMount{
-							{
-								Name:      createVolumeName,
-								MountPath: path.Join(volumeMountPath, "create"),
-								ReadOnly:  true,
-							},
-						},
-					},
-				},
-				RestartPolicy: v1.RestartPolicyNever,
-			},
-		}
-		By("Creating the pod")
-		f.PodClient().CreateSync(pod)
-
-		pollCreateLogs := func() (string, error) {
-			return framework.GetPodLogs(f.ClientSet, f.Namespace.Name, pod.Name, createContainerName)
-		}
-		Eventually(pollCreateLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("Error reading file /etc/secret-volumes/create/data-1"))
-
-		pollUpdateLogs := func() (string, error) {
-			return framework.GetPodLogs(f.ClientSet, f.Namespace.Name, pod.Name, updateContainerName)
-		}
-		Eventually(pollUpdateLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("Error reading file /etc/secret-volumes/update/data-3"))
-
-		pollDeleteLogs := func() (string, error) {
-			return framework.GetPodLogs(f.ClientSet, f.Namespace.Name, pod.Name, deleteContainerName)
-		}
-		Eventually(pollDeleteLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("value-1"))
-
-		By(fmt.Sprintf("Deleting secret %v", deleteSecret.Name))
-		err = f.ClientSet.Core().Secrets(f.Namespace.Name).Delete(deleteSecret.Name, &metav1.DeleteOptions{})
-		Expect(err).NotTo(HaveOccurred(), "Failed to delete secret %q in namespace %q", deleteSecret.Name, f.Namespace.Name)
-
-		By(fmt.Sprintf("Updating secret %v", updateSecret.Name))
-		updateSecret.ResourceVersion = "" // to force update
-		delete(updateSecret.Data, "data-1")
-		updateSecret.Data["data-3"] = []byte("value-3")
-		_, err = f.ClientSet.Core().Secrets(f.Namespace.Name).Update(updateSecret)
-		Expect(err).NotTo(HaveOccurred(), "Failed to update secret %q in namespace %q", updateSecret.Name, f.Namespace.Name)
-
-		By(fmt.Sprintf("Creating secret with name %s", createSecret.Name))
-		if createSecret, err = f.ClientSet.Core().Secrets(f.Namespace.Name).Create(createSecret); err != nil {
-			framework.Failf("unable to create test secret %s: %v", createSecret.Name, err)
-		}
-
-		By("waiting to observe update in volume")
-
-		Eventually(pollCreateLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("value-1"))
-		Eventually(pollUpdateLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("value-3"))
-		Eventually(pollDeleteLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("Error reading file /etc/secret-volumes/delete/data-1"))
-	})
 
 	It("should be consumable from pods in env vars [Conformance]", func() {
 		name := "secret-test-" + string(uuid.NewUUID())
